+++ WoWAPI/connected_realms.py
@@ -2,13 +2,14 @@ from WoWAPI import connector
 
 
 class ConnectedRealms:
-    def __init__(self, region, token):
+    def __init__(self, region, token, namespace):
         self.region = region
         self.token = token
+        self.namespace = namespace
         self.connected_realms = self.get_connected_realms()
 
     def get_connected_realms(self):
-        return connector.Connect(self.region, '/data/wow/connected-realm/index?locale=en_GB', self.token).get_data()
+        return connector.Connect(self.region, '/data/wow/connected-realm/index?locale=en_GB', self.token, self.namespace).get_data()
 
     def get_connected_realms_id(self):
         data = self.connected_realms['connected_realms']
@@ -21,4 +22,4 @@ class ConnectedRealms:
 
     def get_connected_realm(self, realm_id):
         return connector.Connect(self.region, '/data/wow/connected-realm/{}?locale=en_GB'.format(realm_id),
-                                 self.token).get_data()
+                                 self.token, self.namespace).get_data()