+++ WoWAPI/connector.py
@@ -2,16 +2,17 @@ import requests
 
 
 class Connect:
-    def __init__(self, region, endpoint, token, **fields):
+    def __init__(self, region, endpoint, token, namespace, **fields):
         self.region = region
         self.endpoint = endpoint
         self.token = token
         self.fields = fields
+        self.namespace = namespace
 
     def get_data(self):
         url = 'https://{}.api.blizzard.com{}'.format(self.region, self.endpoint)
         payload = {
-            'namespace': 'dynamic-{}'.format(self.region),
+            'namespace': self.namespace,
             'access_token': self.token
         }
         r = requests.get(url, headers=self.token, params={**payload, **self.fields})